sz同士の掛け算でオーバーフローしたので

+++ ___template.cpp
@@ -22,7 +22,7 @@ template<class T> using dijk = priority_queue<T, vector<T>, greater<T>>;
 # define UPPER(rprpr)         transform(all((rprpr)),(rprpr).begin(),TU<char>)
 # define rep(i,upupu)         for(ll i = 0, i##_len = (upupu);(i) < (i##_len);(i)++)
 # define reps(i,opopo)        for(ll i = 1, i##_len = (opopo);(i) <= (i##_len);(i)++)
-# define len(x)                ((int)(x).size())
+# define len(x)                ((ll)(x).size())
 # define bit(n)               (1LL << (n))
 # define pb push_back
 # define exists(c, e)         ((c).find(e) != (c).end())
